sigma_backend/backends/opensearch_ppl/opensearch_ppl.py
"""
OpenSearch PPL backend for Sigma rules.

This backend converts Sigma detection rules into PPL (Piped Processing Language)
queries for OpenSearch.
"""
from typing import List, Optional
from sigma.collection import SigmaCollection
from sigma.rule import SigmaRule


class OpenSearchPPLBackend:
    """
    Backend for converting Sigma rules to OpenSearch PPL queries.
    """
    
    def __init__(self, processing_pipeline: Optional[object] = None):
        """
        Initialize the OpenSearch PPL backend.
        
        Args:
            processing_pipeline: Optional processing pipeline for rule transformation
        """
        self.processing_pipeline = processing_pipeline
    
    def convert(self, sigma_collection: SigmaCollection) -> str:
        """
        Convert a Sigma collection to PPL query.
        
        Args:
            sigma_collection: Collection of Sigma rules to convert
            
        Returns:
            PPL query string
        """
        if not sigma_collection:
            return ""
        
        # For now, return a placeholder implementation
        # This will be replaced with actual PPL conversion logic
        ppl_queries = []
        
        for rule in sigma_collection:
            ppl_query = self.convert_rule(rule)
            if ppl_query:
                ppl_queries.append(ppl_query)
        
        # Join multiple queries if needed
        if len(ppl_queries) == 1:
            return ppl_queries[0]
        elif len(ppl_queries) > 1:
            # Combine multiple queries (adjust based on PPL syntax)
            return " | ".join(ppl_queries)
        else:
            return ""
    
    def convert_rule(self, rule: SigmaRule) -> str:
        """
        Convert a single Sigma rule to PPL query.
        
        Args:
            rule: Single Sigma rule to convert
            
        Returns:
            PPL query string for the rule
        """
        # Placeholder implementation
        # TODO: Implement actual PPL conversion logic
        return f"source = * | where true"